+++ android/versioned-abis/expoview-abi32_0_0/src/main/java/abi32_0_0/host/exp/exponent/modules/api/components/svg/UseView.java
@@ -87,7 +87,7 @@ class UseView extends RenderableView {
 
         float[] dst = new float[2];
         mInvMatrix.mapPoints(dst, src);
-        mInvTransform.mapPoints(dst, src);
+        mInvTransform.mapPoints(dst);
 
         VirtualView template = getSvgView().getDefinedTemplate(mHref);
         int hitChild = template.hitTest(dst);