+++ ios/versioned-react-native/ABI32_0_0/Expo/Core/Api/Components/Svg/Text/ABI32_0_0RNSVGText.m
@@ -135,7 +135,7 @@
     [self renderGroupTo:context rect:rect];
     CGContextRestoreGState(context);
 
-    CGPathRef transformedPath = CGPathCreateCopyByTransformingPath(path, &CGAffineTransformIdentity);
+    CGPathRef transformedPath = CGPathCreateCopy(path);
     [self setHitArea:transformedPath];
     CGPathRelease(transformedPath);
 }